+++ src/battle_anim_effects_3.c
@@ -1692,7 +1692,7 @@ void AnimTask_RapinSpinMonElevation(u8 taskId)
     s16 var2;
     int var3;
     int var4;
-    s16 i;
+    s32 i;
     struct ScanlineEffectParams scanlineParams;
     struct Task *task = &gTasks[taskId];
 
@@ -1768,7 +1768,7 @@ void AnimTask_RapinSpinMonElevation(u8 taskId)
 
 static void RapinSpinMonElevation_Step(u8 taskId)
 {
-    s16 i;
+    s32 i;
     struct Task *task = &gTasks[taskId];
 
     task->data[0] -= task->data[5];
@@ -3275,7 +3275,7 @@ static void AnimTask_AcidArmorStep(u8 taskId)
     s16 bgX, bgY;
     s16 offset;
     s16 var0;
-    s16 i;
+    s32 i;
     s16 sineIndex;
     s16 var3;
 
@@ -3693,7 +3693,7 @@ static void AnimTask_SquishAndSweatDropletsStep(u8 taskId)
 
 static void CreateSweatDroplets(u8 taskId, bool8 arg1)
 {
-    u8 i;
+    u32 i;
     s8 xOffset, yOffset;
     struct Task *task;
     s16 xCoords[4];
@@ -3850,7 +3850,7 @@ void AnimTask_GlareEyeDots(u8 taskId)
 
 static void AnimTask_GlareEyeDotsStep(u8 taskId)
 {
-    u8 i;
+    u32 i;
     s16 x, y;
     struct Task *task = &gTasks[taskId];
