+++ src/battle_anim_mons.c
@@ -770,7 +770,7 @@ u8 GetBattlerPosition(u8 battlerId)
 
 u8 GetBattlerAtPosition(u8 position)
 {
-    u8 i;
+    u32 i;
 
     for (i = 0; i < gBattlersCount; ++i)
         if (gBattlerPositions[i] == position)
@@ -1441,7 +1441,7 @@ void AnimSnoreZ(struct Sprite *sprite)
 
 s16 CloneBattlerSpriteWithBlend(u8 animBattler)
 {
-    u16 i;
+    u32 i;
     u8 spriteId = GetAnimBattlerSpriteId(animBattler);
 
     if (spriteId != 0xFF)
@@ -1711,7 +1711,7 @@ static u16 GetBattlerYDeltaFromSpriteId(u8 spriteId)
     struct BattleSpriteInfo *spriteInfo;
     u8 battlerId = gSprites[spriteId].data[0];
     u16 species;
-    u16 i;
+    u32 i;
 
     for (i = 0; i < MAX_BATTLERS_COUNT; ++i)
     {
