+++ src/battle_anim_special.c
@@ -1517,7 +1517,7 @@ static void PokeBallOpenParticleAnimation_Step2(struct Sprite *sprite)
 
 static void TimerBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1551,7 +1551,7 @@ static void TimerBallOpenParticleAnimation(u8 taskId)
 
 static void DiveBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1586,7 +1586,7 @@ static void DiveBallOpenParticleAnimation(u8 taskId)
 // Also used for Net Ball
 static void SafariBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1621,7 +1621,7 @@ static void SafariBallOpenParticleAnimation(u8 taskId)
 // Also used for Nest Ball
 static void UltraBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1656,7 +1656,7 @@ static void UltraBallOpenParticleAnimation(u8 taskId)
 // Also used for Luxury Ball
 static void GreatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1712,7 +1712,7 @@ static void FanOutBallOpenParticles_Step1(struct Sprite *sprite)
 
 static void RepeatBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
 
@@ -1800,7 +1800,7 @@ static void MasterBallOpenParticleAnimation(u8 taskId)
 
 static void PremierBallOpenParticleAnimation(u8 taskId)
 {
-    u8 i;
+    u32 i;
     u8 x, y, priority, subpriority, ballId;
     u8 spriteId;
