+++ src/cable_car_util.c
@@ -2,8 +2,8 @@
 
 static void CableCarUtil_FillWrapped(void *dest, u16 value, u8 left, u8 top, u8 width, u8 height)
 {
-    u8 i;
-    u8 j;
+    u32 i;
+    u32 j;
     u8 x;
     u8 y;
 
@@ -20,8 +20,8 @@ static void CableCarUtil_FillWrapped(void *dest, u16 value, u8 left, u8 top, u8 
 
 static void CableCarUtil_CopyWrapped(void *dest, const u16 *src, u8 left, u8 top, u8 width, u8 height)
 {
-    u8 i;
-    u8 j;
+    u32 i;
+    u32 j;
     u8 x;
     u8 y;
     const u16 *_src;
